move preferences from a page to a modal
